requirements.txt
python-telegram-bot
opencv-python
dotenv
Pillow